cdk-project/requirements.txt
constructs>=10.0.0,<11.0.0
aws-cdk-lib==2.0.0

